Sampling rate of amps in OBCI was set as a int instead of Float. Normalized all sampling rates of amps in svarog as double - to reflect python float refs: #34246

+++ svarog/src/main/java/org/signalml/app/view/document/opensignal/elements/SignalParametersPanel.java
@@ -338,7 +338,7 @@ public class SignalParametersPanel extends JPanel {
 				getSampleTypeComboBox().setSelectedItem(rawSignalDescriptor.getSampleType());
 		} else if (openSignalDescriptor instanceof ExperimentDescriptor) {
 			ExperimentDescriptor experimentDescriptor = (ExperimentDescriptor) openSignalDescriptor;
-			List<Float> samplingFrequencies = experimentDescriptor.getAmplifier().getSamplingFrequencies();
+			List<Double> samplingFrequencies = experimentDescriptor.getAmplifier().getSamplingFrequencies();
 			getSamplingFrequencyComboBox().setModel(new DefaultComboBoxModel(samplingFrequencies.toArray()));
 
 			float samplingFrequency = experimentDescriptor.getSignalParameters().getSamplingFrequency();
